refactor(activitykit): consolidate PrayerActivityAttributes into single source
Previously two drifted copies existed:
- IqamahLiveActivity/PrayerActivityAttributes.swift (public, init-based)
- iqamah/iOS/PrayerActivityAttributes.swift (internal, member-init, #if iOS)

The iOS-side comment claimed they "must be identical… so ActivityKit can
match them at runtime" — but they weren't. Any future field added to one and
forgotten on the other would silently break Live Activity matching.

Fix: single file at IqamahLiveActivity/PrayerActivityAttributes.swift,
multi-target membership in pbxproj so both iqamah-iOS app and
IqamahLiveActivity extension compile the same definition. Internal access
(default) is sufficient since each target compiles its own module copy.

Both iqamah-iOS and IqamahLiveActivity schemes build clean (iPhone 17 sim).
